requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
drf-yasg
django-filter
django-taggit
pillow